+++ src/components/Footer/FooterMain/FooterMain.tsx
@@ -10,7 +10,7 @@ export default function FooterMain() {
             title="About me"
             content="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed pellentesque nibh enim, quis euismod enim lacinia nec. Phasellus quam diam, semper in erat eu, efficitur molestie purus. Sed a elementum mi. Sed interdum mattis risus, sit amet eleifend ligula luctus ut."
           />
-          <PostsWidget title="Latest posts" limit={3} style="small" />
+          <PostsWidget title="Latest posts" limit={3} style="small" location="footer" />
           <ProductsWidget title="Latest products" limit={2} style="small" />
         </div>
       </Container>