+++ src/components/Widgets/Posts/usePosts.ts
@@ -1,23 +1,13 @@
-import { useState, useEffect } from 'react';
+import { useQuery } from '@tanstack/react-query';
 
 import { getPosts } from '@server/posts/getPosts';
 import type { Post } from '@typings/posts/post';
 
 export default function usePosts(limit: number, category?: number) {
-  const [posts, setPosts] = useState<Post[]>([]);
+  const postsWidgetQuery = useQuery<Post[] | null>({
+    queryKey: ['postsWidget', limit, category],
+    queryFn: () => getPosts(limit, category),
+  });
 
-  useEffect(() => {
-    const fetchPosts = async () => {
-      try {
-        const postsList = await getPosts(limit, category);
-        if (postsList) setPosts(postsList);
-      } catch (error) {
-        console.error('Failed to fetch posts.', error);
-      }
-    };
-
-    fetchPosts();
-  }, [limit, category]);
-
-  return posts;
+  return postsWidgetQuery;
 }